refactor(dtos): change #dtos to export from an `index.ts` and move `GuildQueryType`
