+++ apps/api/src/guild/guild-query.enum.ts
@@ -1,0 +1,5 @@
+export enum GuildQuery {
+  ID = 'ID',
+  NAME = 'NAME',
+  PLAYER = 'PLAYER',
+}
