+++ apps/api/src/guild/guild.controller.ts
@@ -1,4 +1,4 @@
-import { GuildDto } from '#dtos/guild.dto';
+import { GuildDto } from '#dtos';
 import { ErrorResponse, GetGuildResponse } from '#responses';
 import { Controller, Get, Query } from '@nestjs/common';
 import { ApiBadRequestResponse, ApiOkResponse, ApiOperation } from '@nestjs/swagger';
